fix: do not drop dynamic tables when query changes
Previously, there was a `or_replace` attribute on the
snowflake_dynamic_table resource that was meant to specify whether
`OR REPLACE` is meant to be used if the table already exists. However,
this attribute was confusing to use because it often resulted in
spurious changes in terraform plans after applying.

A separate issue with dynamic tables was that we have been dropping and
recreating dynamic tables when the `query` or `refresh_mode` properties
changed, when we could take a less destructive action by calling `CREATE
OR REPLACE`. The presence of `or_replace` as a configured attribute made
this more complicated.

This commit removes the ability for users to set `or_replace` on dynamic
table configurations, and instead will always use `CREATE OR REPLACE`
for dynamic table creation. With this simplification, we can then check
for changes to the `query` and `refresh_mode` arguments when updating
the dynamic table, and either user `CREATE OR REPLACE` or ALTER DYNAMIC
TABLE` depending on which attributes have changed. This allows users to
update the query and refresh mode without causing temporary
unavailability.